requirements.txt
feedparser==6.0.11
google-generativeai==0.8.3
arxiv==2.1.3
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.4
beautifulsoup4==4.12.3
lxml==5.3.0
pywin32==308
